refactor: rename WriteBuffer directory to ScreenBuffer
